backend/requirements.txt
grpcio
grpcio-tools
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]  # For JWT
passlib
dotenv